+++ Tweaks/SearchTheMarkets.cs
@@ -14,7 +14,16 @@ namespace HaselTweaks.Tweaks;
 
 [Tweak(
     Name: "Search the markets",
-    Description: "Adds an entry to item context menus that allows you to quickly search for the item on the market board. Only visible when market board is open!\n\nSupported context menus are:\n- Chat\n- Crafting Log\n- Grand Company Delivery Missions\n- Inventory\n- Materials List\n- Recipe Tree"
+    Description: @"Adds an entry to item context menus that allows you to quickly search for the item on the market board. Only visible when market board is open!
+
+Supports context menus in the following windows:
+- Chat
+- Crafting Log
+- Ehcatl Nine Delivery Quests (via /timers)
+- Grand Company Delivery Missions (via /timers)
+- Inventory
+- Materials List
+- Recipe Tree"
 )]
 public unsafe class SearchTheMarkets : Tweak
 {
@@ -77,7 +86,7 @@ public unsafe class SearchTheMarkets : Tweak
 
     private void ContextMenu_OnOpenGameObjectContextMenu(GameObjectContextMenuOpenArgs args)
     {
-        if (args.ParentAddonName is not ("RecipeNote" or "RecipeMaterialList" or "RecipeTree" or "ChatLog" or "ContentsInfoDetail"))
+        if (args.ParentAddonName is not ("RecipeNote" or "RecipeMaterialList" or "RecipeTree" or "ChatLog" or "ContentsInfoDetail" or "DailyQuestSupply"))
             return;
 
         _itemId = 0u;
@@ -104,6 +113,11 @@ public unsafe class SearchTheMarkets : Tweak
             case "ContentsInfoDetail":
                 if (GetAgent<AgentContentsTimer>(AgentId.ContentsTimer, out var agentContentsTimer))
                     _itemId = agentContentsTimer->ContextMenuItemId;
+                break;
+
+            case "DailyQuestSupply":
+                if (GetAgent<AgentDailyQuestSupply>(AgentId.DailyQuestSupply, out var agentDailyQuestSupply))
+                    _itemId = agentDailyQuestSupply->ContextMenuItemId;
                 break;
         }